refactor(key) : 키분리
기존에 jwt config로 관리되던 API키를 분리하였습니다.
